alteração da arquitetura do projeto
