+++ cliente-pacote-tarifas/src/main/java/com/brq/cliente/pacote/tarifas/ClientePacoteTarifasApplication.java
@@ -1,0 +1,13 @@
+package com.brq.cliente.pacote.tarifas;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ClientePacoteTarifasApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ClientePacoteTarifasApplication.class, args);
+	}
+
+}
