+++ cliente-pacote-tarifas/src/main/java/com/brq/cliente/pacote/tarifas/repositories/ClienteRepository.java
@@ -1,0 +1,11 @@
+package com.brq.cliente.pacote.tarifas.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.brq.cliente.pacote.tarifas.entities.Cliente;
+
+@Repository
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+
+}
