+++ cliente-pacote-tarifas/src/main/java/com/brq/cliente/pacote/tarifas/services/ClienteService.java
@@ -1,0 +1,66 @@
+package com.brq.cliente.pacote.tarifas.services;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Service;
+
+import com.brq.cliente.pacote.tarifas.entities.Cliente;
+import com.brq.cliente.pacote.tarifas.repositories.ClienteRepository;
+import com.brq.cliente.pacote.tarifas.services.exception.DataIntegrityException;
+import com.brq.cliente.pacote.tarifas.services.exception.ObjectNotFoundException;
+
+
+
+@Service
+public class ClienteService {
+	
+	@Autowired
+	private ClienteRepository repository;
+	
+
+	public List<Cliente> findAll() {
+		return  repository.findAll();
+	}
+	
+	public Cliente insert(Cliente obj) {
+		obj.setId(null);
+		obj = repository.save(obj);
+		return obj;
+	}
+
+	public Cliente update(Cliente obj) {
+		Cliente newObj = find(obj.getId());
+		upDateDate(newObj, obj);
+		return repository.save(newObj);
+	}
+	
+	public void delete(Long id) {
+		find(id);
+		try {
+			repository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DataIntegrityException("Não é possivel excluir o cliente");
+			
+		}
+	}
+	
+	
+	
+	public Cliente find(Long id) {
+		Optional<Cliente> obj = repository.findById(id);
+		return obj.orElseThrow(() -> new ObjectNotFoundException("Objeto não encontrado! Id: "+id+" Tipo: "+Cliente.class.getName()));
+		
+	}
+	
+	
+	private void upDateDate(Cliente newObj, Cliente obj) {
+		newObj.setNome(obj.getNome());
+		
+	}
+
+	
+
+}
