+++ cliente-pacote-tarifas/src/main/java/com/brq/cliente/pacote/tarifas/services/exception/DataIntegrityException.java
@@ -1,0 +1,15 @@
+package com.brq.cliente.pacote.tarifas.services.exception;
+
+public class DataIntegrityException extends RuntimeException {
+
+	private static final long serialVersionUID = 1L;
+	
+	public DataIntegrityException(String msg) {
+		super(msg);
+	}
+	
+	public DataIntegrityException(String msg, Throwable cause) {
+		super(msg, cause);
+	}
+
+}
