+++ cliente-pacote-tarifas/src/main/java/com/brq/cliente/pacote/tarifas/services/exception/ObjectNotFoundException.java
@@ -1,0 +1,15 @@
+package com.brq.cliente.pacote.tarifas.services.exception;
+
+public class ObjectNotFoundException extends RuntimeException {
+
+	private static final long serialVersionUID = 1L;
+	
+	public ObjectNotFoundException(String msg) {
+		super(msg);
+	}
+	
+	public ObjectNotFoundException(String msg, Throwable cause) {
+		super(msg, cause);
+	}
+
+}
