+++ cliente-pacote-tarifas/src/main/resources/application.properties
@@ -1,0 +1,13 @@
+
+spring.application.name=cliente
+server.port=8080
+
+# Database configuration
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.username=sa
+spring.datasource.password=
+
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
+
+spring.jpa.defer-datasource-initialization=true
