+++ cliente-pacote-tarifas/src/main/resources/data.sql
