+++ cliente-pacote-tarifas/src/test/java/com/brq/cliente/pacote/tarifas/ClientePacoteTarifasApplicationTests.java
@@ -1,0 +1,13 @@
+package com.brq.cliente.pacote.tarifas;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ClientePacoteTarifasApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
