--- cliente/src/main/java/com/brq/cliente/ClienteApplication.java
@@ -1,13 +1,0 @@
-package com.brq.cliente;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ClienteApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ClienteApplication.class, args);
-	}
-
-}
