--- cliente/src/main/java/com/brq/cliente/repositories/ClienteRepository.java
@@ -1,11 +1,0 @@
-package com.brq.cliente.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.brq.cliente.entities.Cliente;
-
-@Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
-
-}
