--- cliente/src/main/java/com/brq/cliente/service/exception/ObjectNotFoundException.java
@@ -1,15 +1,0 @@
-package com.brq.cliente.service.exception;
-
-public class ObjectNotFoundException extends RuntimeException {
-
-	private static final long serialVersionUID = 1L;
-	
-	public ObjectNotFoundException(String msg) {
-		super(msg);
-	}
-	
-	public ObjectNotFoundException(String msg, Throwable cause) {
-		super(msg, cause);
-	}
-
-}
