--- cliente/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-spring.application.name=cliente
-server.port=8001
-
-# Database configuration
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.username=sa
-spring.datasource.password=
-
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
-
-spring.jpa.defer-datasource-initialization=true
