--- cliente/src/test/java/com/brq/cliente/ClienteApplicationTests.java
@@ -1,13 +1,0 @@
-package com.brq.cliente;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ClienteApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
