--- pacote-tarifas/src/main/java/com/brq/pacote/tarifas/PacoteTarifasApplication.java
@@ -1,13 +1,0 @@
-package com.brq.pacote.tarifas;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PacoteTarifasApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PacoteTarifasApplication.class, args);
-	}
-
-}
