--- pacote-tarifas/src/main/java/com/brq/pacote/tarifas/entities/PacoteTarifa.java
@@ -1,92 +1,0 @@
-package com.brq.pacote.tarifas.entities;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-
-@Entity
-@Table(name = "tb_pacoteTarifa")
-public class PacoteTarifa implements Serializable {
-	private static final long serialVersionUID = 1L;
-	
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	private String nome;
-	private Double valorMinimo;
-	private Double valorMaximo;
-	
-	public PacoteTarifa() {
-		
-	}
-	
-	public PacoteTarifa(Long id, String nome, Double valorMinimo, Double valorMaximo) {
-		super();
-		this.id = id;
-		this.nome = nome;
-		this.valorMinimo = valorMinimo;
-		this.valorMaximo = valorMaximo;
-	}
-
-
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public Double getValorMinimo() {
-		return valorMinimo;
-	}
-
-	public void setValorMinimo(Double valorMinimo) {
-		this.valorMinimo = valorMinimo;
-	}
-
-	public Double getValorMaximo() {
-		return valorMaximo;
-	}
-
-	public void setValorMaximo(Double valorMaximo) {
-		this.valorMaximo = valorMaximo;
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(id);
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		PacoteTarifa other = (PacoteTarifa) obj;
-		return Objects.equals(id, other.id);
-	}
-	
-	
-	
-	
-
-}
