--- pacote-tarifas/src/main/java/com/brq/pacote/tarifas/service/exception/DataIntegrityException.java
@@ -1,15 +1,0 @@
-package com.brq.pacote.tarifas.service.exception;
-
-public class DataIntegrityException extends RuntimeException {
-
-	private static final long serialVersionUID = 1L;
-	
-	public DataIntegrityException(String msg) {
-		super(msg);
-	}
-	
-	public DataIntegrityException(String msg, Throwable cause) {
-		super(msg, cause);
-	}
-
-}
