--- pacote-tarifas/src/main/resources/data.sql
@@ -1,4 +1,0 @@
-
-INSERT INTO tb_pacote_Tarifa (nome, valor_Minimo, valor_Maximo) VALUES ('Padronizado I', 30.0, 39.9);
-INSERT INTO tb_pacote_Tarifa (nome, valor_Minimo, valor_Maximo) VALUES ('Padronizado II', 40.0, 49.9);
-INSERT INTO tb_pacote_Tarifa (nome, valor_Minimo, valor_Maximo) VALUES ('Padronizado III', 50.0, 59.9);
