--- pacote-tarifas/src/test/java/com/brq/pacote/tarifas/PacoteTarifasApplicationTests.java
@@ -1,13 +1,0 @@
-package com.brq.pacote.tarifas;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PacoteTarifasApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
